debug homography

+++ cpp_cuda/src/faceswap_func_cuda/pybind_cuda.cuh
@@ -4,11 +4,11 @@
 extern "C"{
     void gpu_square_main(int * v_in, int size, int * v_out);
     //void ApplyAllHomography(int width, int height, int* imgLabel, double** H, double* XI, double* YI, int n_quadrangles);
-    void ApplyAllHomography_CUDA(int width, int height, int* imgLabel, double* H_flat, int* XI, int* YI, int n_quadrangles); 
+    void ApplyAllHomography_CUDA(int width, int height, int* imgLabel, double* H_flat, double* XI, double* YI, int n_quadrangles); 
     void RecreateImage_CUDA(int* imgOut,
 	                        int* imgCAM, int width_CAM, int height_CAM,
 	                        int* imgFTA, int width_FTA, int height_FTA,
- 	                        int* XI, int* YI, int* imgLabel);
+ 	                        double* XI, double* YI, int* imgLabel);
 }
